// src/store/main/system/type.ts
interface IUser {
  id: number;
  name: string;
  realname: string;
  cellphone: number;
  enable: number;
  departmentId: number;
  roleId: number;
  createAt: string;
  updateAt: string;
}

export interface ISystemState {
  // json to typescript网站
  usersList: IUser[];
  usersTotalCount: number;

  pageList: any[];
  pageTotalCount: number;
}